Make ssize_t clean

+++ Modules/_sqlite/connection.c
@@ -377,7 +377,7 @@ void _set_result(sqlite3_context* context, PyObject* py_val)
 {
     long longval;
     const char* buffer;
-    int buflen;
+    Py_ssize_t buflen;
     PyObject* stringval;
 
     if (PyErr_Occurred()) {
@@ -416,7 +416,7 @@ PyObject* _build_py_params(sqlite3_context *context, int argc, sqlite3_value** a
     PyObject* cur_py_value;
     const char* val_str;
     PY_LONG_LONG val_int;
-    int buflen;
+    Py_ssize_t buflen;
     void* raw_buffer;
 
     args = PyTuple_New(argc);
